requirements.txt
# QBitaLabs - Core Dependencies
# Install with: pip install -r requirements.txt

# Core
numpy>=1.24
scipy>=1.10
pandas>=2.0

# ML/AI
torch>=2.0
transformers>=4.30
torch-geometric>=2.3

# Agents
langchain>=0.1
langgraph>=0.0.30
anthropic>=0.25

# Quantum - Qiskit (IBM)
qiskit>=1.0
qiskit-ibm-runtime>=0.20
qiskit-nature>=0.7
qiskit-algorithms>=0.3

# Quantum - Cirq (Google)
cirq>=1.3
cirq-google>=1.3
openfermion>=1.6
openfermionpyscf>=0.5

# Quantum - PennyLane
pennylane>=0.35
pennylane-qiskit>=0.35

# Chemistry
rdkit>=2023.3
biopython>=1.81
pyscf>=2.3

# Neuromorphic (simulator)
brian2>=2.5

# API
fastapi>=0.100
uvicorn>=0.22
pydantic>=2.0
websockets>=11.0

# Data
pyarrow>=12.0
h5py>=3.8

# Utils
pyyaml>=6.0
python-dotenv>=1.0
rich>=13.0
typer>=0.9
structlog>=23.1
tenacity>=8.2
